Move utils to common dir
